new folder:   hrl_dynamic_mpc
These files are the updated version of the dynamic MPC controller for
reaching in clutter described in Marc Killpack's dissertation for a 7
degree of freedom robot arm. There are some minor changes since it was
run on the real robot DARCI and there is probably some minor testing
and debugging that is necessary. This folder also contains the matlab
files used to generate simulataion results in the dissertation.

modified:   hrl_haptic_mpc

This includes any updates to hrl_haptic_mpc since the last release
over a year ago. The main purpose of this is to include rviz
teloperation functionality for DARCI. However, it also should be
tested and debugged.